+++ lib/libtpe/tpe/task/ont_block_decompression_task.cpp
@@ -177,7 +177,7 @@ size_t OntBlockDecompressionTask::_dispatch(
     size_t output_offset = offset - BxOntBlockMessage_t::offset_diff;
     size_t prev_output_offset = output_offset;
     bool is_short;
-    for (int count = 0 ; count < _tx_count ; ++count) {
+    for (uint64_t count = 0 ; count < _tx_count ; ++count) {
         size_t from = offset;
         auto& tdata = _sub_tasks[idx]->task_data();
         idx = std::min((size_t) (count / bulk_size), pool_size - 1);
